configure: separate LDFLAGS and extra-libraries fields
This correctly addresses robeverest/cufft#4

The problem with the combined LDFLAGS macro is that this is expanded too early on the command line. Namely, we have -lcufft placed before the object files that use these symbols. Since the linker has not yet encountered any uses of cufft symbols at that time, in incorrectly assumes that it is not used and strips all the symbols.

See also tmcdonell/cuda#13