add 'setStream'
All operations in the given CUFFT plan will then occur in that stream, until the stream is changed or the plan destroyed.

This bumps the version requirement on the CUDA package to 0.6.6, which is when the Stream and Event datatypes were shared between the bindings to the runtime and driver APIs. However, the datatypes were actually interchangeable since CUDA-3.1, so it would be possible to (with a small amount of guilt) enable the use of old versions of the package.